@@ -13,33 +13,33 @@
     },
     "private": false,
     "dependencies": {
-        "@emotion/react": "^11.10.5",
-        "@emotion/styled": "^11.10.5",
-        "@mui/icons-material": "^5.10.9",
-        "@mui/material": "^5.10.12",
-        "flo-bezier3": "^4.0.1",
-        "flo-boolean": "^2.1.1",
-        "flo-draw": "^2.0.0",
-        "flo-mat": "^2.0.1",
-        "flo-poly": "^6.0.4",
-        "flo-vector2d": "^4.0.1",
-        "ramda": "^0.28.0",
+        "@emotion/react": "^11.11.1",
+        "@emotion/styled": "^11.11.0",
+        "@mui/icons-material": "^5.14.3",
+        "@mui/material": "^5.14.3",
+        "flo-bezier3": "^5.0.0",
+        "flo-boolean": "^2.1.8",
+        "flo-draw": "^2.0.2",
+        "flo-mat": "^3.0.0",
+        "flo-poly": "^6.0.8",
+        "flo-vector2d": "^4.0.2",
+        "ramda": "^0.29.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "react-router-dom": "^6.4.2"
+        "react-router-dom": "^6.14.2"
     },
     "devDependencies": {
-        "@types/react": "^18.0.24",
-        "@types/react-dom": "^18.0.8",
-        "@types/node": "^18.11.8",
-        "@types/ramda": "^0.28.18",
+        "@types/node": "^20.4.8",
+        "@types/ramda": "^0.29.3",
+        "@types/react": "^18.2.18",
+        "@types/react-dom": "^18.2.7",
         "@types/react-router-dom": "^5.3.3",
-        "resolve-typescript-plugin": "^1.2.0",
+        "resolve-typescript-plugin": "^2.0.1",
         "source-map-loader": "^4.0.1",
-        "ts-loader": "^9.4.1",
-        "typescript": "^4.8.4",
-        "webpack": "^5.74.0",
-        "webpack-cli": "^4.10.0"
+        "ts-loader": "^9.4.4",
+        "typescript": "^5.1.6",
+        "webpack": "^5.88.2",
+        "webpack-cli": "^5.1.4"
     },
     "scripts": {
         "build": "webpack"
